refactor(PathFinder, PathService, PathResponse, PathFinderTest, PathAcceptanceTest): don't read DB in constructor, change return type of getPath(), and map the response body to PathResponse
