+++ src/main/java/nextstep/subway/path/domain/PathFinder.java
@@ -2,11 +2,13 @@ package nextstep.subway.path.domain;
 
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
+import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.domain.Station;
 import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -15,22 +17,26 @@ import java.util.List;
 public class PathFinder {
 
     private WeightedMultigraph<Station, DefaultWeightedEdge> graph = new WeightedMultigraph<>(DefaultWeightedEdge.class);
-    private final LineRepository lineRepository;
+    private LineRepository lineRepository;
 
+    @Autowired
     public PathFinder(LineRepository lineRepository) {
         this.lineRepository = lineRepository;
-        renewGraph();
+    }
+
+    public PathFinder(List<Line> lines) {
+        enrollLinesInGraph(lines);
     }
 
     private void enrollLinesInGraph(List<Line> lines) {
         lines.forEach(line -> line.enrollIn(graph));
     }
 
-    public void addLines(List<Line> lines) {
-        enrollLinesInGraph(lines);
-    }
+    public PathResponse getPath(Station source, Station target) {
+        if (graph.vertexSet().isEmpty()) {
+            renewGraph();
+        }
 
-    public GraphPath<Station, DefaultWeightedEdge> getPath(Station source, Station target) {
         isSame(source, target);
         isExistent(source, target);
 
@@ -40,7 +46,13 @@ public class PathFinder {
             throw new IllegalArgumentException("출발역과 도착역은 연결되어 있어야 합니다.");
         }
 
-        return path;
+        return PathResponse.of(path.getVertexList(), path.getWeight());
+    }
+
+    public void renewGraph() {
+        graph = new WeightedMultigraph<>(DefaultWeightedEdge.class);
+        lineRepository.findAll()
+                .forEach(line -> line.enrollIn(graph));
     }
 
     private void isSame(Station source, Station target) {
@@ -53,11 +65,5 @@ public class PathFinder {
         if (!graph.containsVertex(source) || !graph.containsVertex(target)) {
             throw new IllegalArgumentException("존재하지 않는 역입니다.");
         }
-    }
-
-    public void renewGraph() {
-        graph = new WeightedMultigraph<>(DefaultWeightedEdge.class);
-        lineRepository.findAll()
-                .forEach(line -> line.enrollIn(graph));
     }
 }
